pyproject.toml
[tool.poetry]
name = "bicyle-planner"
version = "0.0.1"
description = "QGIS Bicycle Planner Plugin"
authors = ["Joakim Fors <joakim.fors@trivector.se>"]
license = "GPL-2.0"

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.812"
pytest = "^6.2.3"
flake8 = "^3.9.1"
isort = "^5.8.0"
autoflake = "^1.4"
pytest-qgis = "^0.1.0"

[tool.black]
line-length = 88
skip-string-normalization = true
target-version = ['py37']
include = '\.pyi?$'

[tool.isort]
line_length = 88
multi_line_output = 3
balanced_wrapping = true
combine_as_imports = true
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
minversion = "6.0"
#addopts = "-ra -q"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
